+++ app/src/androidTest/java/ch/epfl/sdp/map/MapActivityTest.java
@@ -9,6 +9,7 @@ import org.junit.Rule;
 import org.junit.Test;
 
 import ch.epfl.sdp.R;
+import ch.epfl.sdp.testActivities.MapActivity;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
