+++ app/src/main/AndroidManifest.xml
@@ -32,7 +32,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".map.MapActivity">
+        <activity android:name=".testActivities.MapActivity">
             <intent-filter>
                 <action android:name="applicationB.intent.action.Launch" />
                 <category android:name="android.intent.category.DEFAULT" />
