+++ app/src/main/java/ch/epfl/sdp/IntroActivity.java
@@ -2,11 +2,8 @@ package ch.epfl.sdp;
 
 import android.graphics.Color;
 import android.os.Bundle;
-import android.view.View;
 
 import com.github.paolorotolo.appintro.AppIntro;
-import com.github.paolorotolo.appintro.AppIntro2;
-import com.github.paolorotolo.appintro.AppIntro2Fragment;
 import com.github.paolorotolo.appintro.AppIntroFragment;
 import com.github.paolorotolo.appintro.model.SliderPage;
 import com.google.android.gms.auth.api.signin.GoogleSignIn;
@@ -14,7 +11,7 @@ import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import ch.epfl.sdp.fragment.AuthenticationFragment;
+import ch.epfl.sdp.identity.fragment.AuthenticationFragment;
 
 public class IntroActivity extends AppIntro {
