+++ app/src/main/java/ch/epfl/sdp/identity/Account.java
@@ -1,4 +1,4 @@
-package ch.epfl.sdp.utilities;
+package ch.epfl.sdp.identity;
 
 import android.app.Activity;
 import android.net.Uri;
