+++ app/src/main/java/ch/epfl/sdp/identity/AccountFactory.java
@@ -1,4 +1,4 @@
-package ch.epfl.sdp.utilities;
+package ch.epfl.sdp.identity;
 
 //import Account;
