+++ app/src/main/java/ch/epfl/sdp/testActivities/AccountActivity.java
@@ -3,7 +3,7 @@ package ch.epfl.sdp.testActivities;
 import androidx.fragment.app.Fragment;
 
 import ch.epfl.sdp.utilities.SingleFragmentActivity;
-import ch.epfl.sdp.fragment.AccountFragment;
+import ch.epfl.sdp.identity.fragment.AccountFragment;
 
 /**
  * Class AccounteGettingactivity : once logged in google, this class will be able to retrieve given user information.
