+++ app/src/main/java/ch/epfl/sdp/testActivities/MapActivity.java
@@ -1,7 +1,8 @@
-package ch.epfl.sdp.map;
+package ch.epfl.sdp.testActivities;
 
 import androidx.fragment.app.Fragment;
 
+import ch.epfl.sdp.map.MapFragment;
 import ch.epfl.sdp.utilities.SingleFragmentActivity;
 
 public class MapActivity extends SingleFragmentActivity {
