+++ app/src/main/java/ch/epfl/sdp/testActivities/UserInfectionActivity.java
@@ -3,7 +3,6 @@ package ch.epfl.sdp.testActivities;
 import androidx.fragment.app.Fragment;
 
 import ch.epfl.sdp.utilities.SingleFragmentActivity;
-import ch.epfl.sdp.fragment.UserInfectionFragment;
 
 public class UserInfectionActivity extends SingleFragmentActivity {
     @Override
