+++ app/src/main/java/ch/epfl/sdp/toDelete/HistoryFirestoreInteractor.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
-import ch.epfl.sdp.utilities.Account;
+import ch.epfl.sdp.identity.Account;
 import ch.epfl.sdp.utilities.PositionRecord;
 import ch.epfl.sdp.firestore.ConcreteFirestoreInteractor;
